# api-contracts/openapi/ctmp.openapi.yaml
openapi: 3.0.3
info:
  title: CTMP API
  version: 1.0.0
  description: >
    Corporate Tender Management Platform API contract. Commercial envelope state
    and commercial visibility are separate controls: committee opening changes
    envelope state only, while commercial details still require explicit
    commercial permissions. All sensitive state changes, views, downloads,
    exports, permission changes, and exceptions must be audit logged.
servers:
  - url: /api/v1
    description: Versioned API root
...
security:
  - bearerAuth: []
paths:
...
  /tenders/{tenderId}/technical-opening:
    post:
      tags: [Technical Evaluation]
      summary: Open technical envelopes
      description: Technical envelopes can open only after Submission Closed. Commercial envelopes remain sealed.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenderId'
      responses:
        '200':
          description: Technical opening completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalOpeningResult'
  /tenders/{tenderId}/technical-evaluations:
    get:
      tags: [Technical Evaluation]
      summary: List technical evaluations for a tender
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenderId'
      responses:
        '200':
          description: Technical evaluations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalEvaluationListResponse'
  /bids/{bidId}/technical-evaluations:
    post:
      tags: [Technical Evaluation]
      summary: Submit technical evaluation for a bid
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BidId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechnicalEvaluationRequest'
      responses:
        '201':
          description: Technical evaluation recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalEvaluation'
  /tenders/{tenderId}/finalize-technical-results:
    post:
      tags: [Technical Evaluation]
      summary: Finalize technical pass/fail results
      description: Only technically passed vendors proceed to commercial comparison.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenderId'
      responses:
        '200':
          description: Technical results finalized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalEvaluationListResponse'
  /tenders/{tenderId}/technical-criteria:
    get:
      tags: [Technical Evaluation]
      summary: Get evaluation criteria configured for tender
      description: Returns criteria list + pass threshold for the scorecard UI. Falls back to department/tender-type defaults if no tender-specific criteria stored.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenderId'
      responses:
        '200':
          description: Criteria configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalCriteriaResponse'
...
  /tenders/{tenderId}/commercial-comparison:
    get:
      tags: [Commercial Evaluation]
      summary: Get commercial comparison
      description: >
        Shows only technically passed vendors. Users without commercial:view see
        status-only rows. Prices, scoring, documents, downloads, and exports
        require explicit commercial permissions and must be audit logged.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenderId'
      responses:
        '200':
          description: Commercial comparison respecting caller permissions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommercialComparison'
  /bids/{bidId}/commercial-evaluations:
    post:
      tags: [Commercial Evaluation]
      summary: Submit commercial evaluation for a bid
      description: Requires envelope OPENED and commercial:evaluate. The action is audit logged.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BidId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommercialEvaluationRequest'
      responses:
        '201':
          description: Commercial evaluation recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommercialEvaluation'
...
  /roles:
    get:
      tags: [Roles]
      summary: List roles
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Role list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'
  /roles/{roleId}/permissions:
    get:
      tags: [Roles]
      summary: Get permission IDs assigned to role
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoleId'
      responses:
        '200':
          description: Permission IDs.
          content:
            application/json:
              schema:
                type: object
                required: [permissionIds]
                properties:
                  permissionIds:
                    type: array
                    items:
                      type: string
                      format: uuid
    patch:
      tags: [Roles]
      summary: Replace permission set for role
      description: System roles (is_system=true) are read-only and return 403. Each add/remove emits an audit event.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permissionIds]
              properties:
                permissionIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Permissions updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  permissionIds:
                    type: array
                    items:
                      type: string
                      format: uuid
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenderId:
      name: tenderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BidId:
      name: bidId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RoleId:
      name: roleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    EnvelopeStatus:
      type: string
      enum: [DRAFT, SUBMITTED, SEALED, OPENED, LOCKED]
...
    TechnicalOpeningResult:
      type: object
      required: [tenderId, openedEnvelopeCount]
      properties:
        tenderId:
          type: string
          format: uuid
        openedEnvelopeCount:
          type: integer
    TechnicalEvaluationRequest:
      type: object
      required: [result]
      properties:
        result:
          type: string
          enum: [PASS, FAIL]
        score:
          type: number
        comments:
          type: string
        scores:
          type: array
          items:
            type: object
            properties:
              criterion:
                type: string
              score:
                type: number
              comments:
                type: string
    TechnicalEvaluation:
      type: object
      required: [id, bidId, result]
      properties:
        id:
          type: string
          format: uuid
        bidId:
          type: string
          format: uuid
        evaluatorUserId:
          type: string
          format: uuid
        result:
          type: string
          enum: [PASS, FAIL]
        score:
          type: number
    TechnicalEvaluationListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/TechnicalEvaluation'
...
    CommercialComparison:
      type: object
      required: [tenderId, rows]
      properties:
        tenderId:
          type: string
          format: uuid
        callerCommercialAccess:
          type: object
          properties:
            canView:
              type: boolean
            canDownload:
              type: boolean
            canEvaluate:
              type: boolean
            canExport:
              type: boolean
        rows:
          type: array
          items:
            $ref: '#/components/schemas/CommercialComparisonRow'
    CommercialComparisonRow:
      type: object
      required: [bidId, vendorId, technicalResult, commercialEnvelopeStatus]
      properties:
        bidId:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
        technicalResult:
          type: string
          enum: [PASS]
        commercialEnvelopeStatus:
          $ref: '#/components/schemas/EnvelopeStatus'
        commercialDetailsVisible:
          type: boolean
        totalAmount:
          type: number
          description: Present only when caller has commercial:view.
        currency:
          type: string
        rank:
          type: integer
    CommercialEvaluationRequest:
      type: object
      required: [score]
      properties:
        score:
          type: number
        comments:
          type: string
    CommercialEvaluation:
      type: object
      required: [id, bidId, score]
      properties:
        id:
          type: string
          format: uuid
        bidId:
          type: string
          format: uuid
        evaluatorUserId:
          type: string
          format: uuid
        score:
          type: number
        comments:
          type: string
...
    TechnicalCriterion:
      type: object
      required: [code, name, maxScore]
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        maxScore:
          type: number
        weight:
          type: number
        mandatory:
          type: boolean
    TechnicalCriteriaResponse:
      type: object
      required: [tenderId, criteria, passThreshold]
      properties:
        tenderId:
          type: string
          format: uuid
        criteria:
          type: array
          items:
            $ref: '#/components/schemas/TechnicalCriterion'
        passThreshold:
          type: number
          description: Minimum total score required for PASS.
        source:
          type: string
          enum: [TENDER, DEPARTMENT_DEFAULT, SYSTEM_DEFAULT]
...
    Role:
      type: object
      required: [id, name, permissionCount]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        permissionCount:
          type: integer
        userCount:
          type: integer
        isSystem:
          type: boolean
    RoleListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Role'
...
    PaginatedResponse:
      type: object
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer